Remove surface magnet option from the testing jig
Realistically, this option should be implemented at the base jig level,
or as a transformation function of any jig.

However, for the testing jig it doesn't make sense in particular to have
them in there - if you approach the PCB from its bottom side, wires will
be sticking out on the same side surface magnets would be. If you
approach it from the top side, the PCB itself will be in the way.